Removed the yuri command and added a yuri command
Instead of saying "hey" it now actually does a thing, wowie!
